refactor(prefer-item): 기관 선호 물품 추가 삭제 리팩토링
