+++ src/main/java/com/somemore/center/dto/response/PreferItemCreateResponseDto.java
@@ -1,14 +1,13 @@
-package com.somemore.domains.center.dto.response;
+package com.somemore.center.dto.response;
 
-import com.fasterxml.jackson.databind.PropertyNamingStrategies;
+import com.fasterxml.jackson.databind.PropertyNamingStrategies.SnakeCaseStrategy;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
-import com.somemore.domains.center.domain.PreferItem;
+import com.somemore.center.domain.PreferItem;
 import io.swagger.v3.oas.annotations.media.Schema;
+import java.util.UUID;
 import lombok.Builder;
 
-import java.util.UUID;
-
-@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
+@JsonNaming(SnakeCaseStrategy.class)
 @Builder
 public record PreferItemCreateResponseDto(
         @Schema(description = "선호물품의 ID", example = "111")
@@ -18,6 +17,7 @@ public record PreferItemCreateResponseDto(
         @Schema(description = "선호물품 이름", example = "어린이 도서")
         String itemName
 ) {
+
     public static PreferItemCreateResponseDto from(PreferItem preferItem) {
         return PreferItemCreateResponseDto.builder()
                 .id(preferItem.getId())
