+++ src/main/java/com/somemore/center/repository/preferitem/PreferItemJpaRepository.java
@@ -1,11 +1,11 @@
-package com.somemore.domains.center.repository.preferitem;
+package com.somemore.center.repository.preferitem;
 
-import com.somemore.domains.center.domain.PreferItem;
-import org.springframework.data.jpa.repository.JpaRepository;
-
+import com.somemore.center.domain.PreferItem;
 import java.util.List;
 import java.util.UUID;
+import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PreferItemJpaRepository extends JpaRepository<PreferItem, Long> {
+
     List<PreferItem> findByCenterId(UUID centerId);
 }
