+++ src/main/java/com/somemore/center/repository/preferitem/PreferItemRepositoryImpl.java
@@ -1,12 +1,11 @@
-package com.somemore.domains.center.repository.preferitem;
+package com.somemore.center.repository.preferitem;
 
-import com.somemore.domains.center.domain.PreferItem;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Repository;
-
+import com.somemore.center.domain.PreferItem;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Repository;
 
 @RequiredArgsConstructor
 @Repository
