+++ src/main/java/com/somemore/center/usecase/DeletePreferItemUseCase.java
@@ -1,7 +1,8 @@
-package com.somemore.domains.center.usecase.command;
+package com.somemore.center.usecase;
 
 import java.util.UUID;
 
 public interface DeletePreferItemUseCase {
+
     void deletePreferItem(UUID centerId, Long preferItemId);
 }
